refactor(providers): move context providers to App Router structure
- Move UserContext to src/app/_providers/UserProvider.tsx to align with Next.js App Router conventions

- Create providers index file with centralized Providers component

- Update imports in all components using the context

- Fix type errors in ProtectedRoute component
