+++ src/app/(main-site)/(auth)/demande-devis-souscription/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useUser } from '@/lib/contexts/UserContext';
+import { useUser } from '@/app/_providers';
 import { Email, Lock, Person, Visibility, VisibilityOff } from '@mui/icons-material';
 import {
   Alert,
